[REFACTOR#1603] 기존 api 옮기면서 필요없어진 내용 정리 및 soptlog domain에 캐싱 전략 추가
