+++ domain/home/src/main/java/org/sopt/official/domain/home/model/UserDescription.kt
@@ -25,5 +25,5 @@
 package org.sopt.official.domain.home.model
 
 data class UserDescription(
-    val activityDescription: String,
+    val activityDescription: String = "",
 )
