+++ domain/soptlog/build.gradle.kts
@@ -32,4 +32,5 @@ kotlin {
 
 dependencies {
     implementation(libs.javax.inject)
+    implementation(libs.kotlin.coroutines)
 }
